Reference commonjs module as main component
Since we are using webpack to build the project and webpack
doesn't currently support ES6 modules we need to chose commonjs
or UMD, we are going with the former.

Incidently as a result this plugin cannot be loaded via the typical
Vue.use(NyanCat) mechanism in main.js, perhaps look into using
rollup to build a ES6 component to mitigate this at some point.